refactor(pkg/version): move version command from cmd to internal/version
